backend/requirements.txt
# Core dependencies
fastapi>=0.95.0
uvicorn>=0.21.1
pydantic>=1.10.7
python-dotenv>=1.0.0
sqlalchemy>=2.0.9
alembic>=1.10.3
psycopg2-binary>=2.9.6
asyncpg>=0.27.0
pgvector>=0.2.0

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.1
python-multipart>=0.0.6

# Document processing
langchain>=0.0.200
pypdf>=3.9.0
python-docx>=0.8.11
beautifulsoup4>=4.12.2
html2text>=2020.1.16
sentence-transformers>=2.2.2

# Agent framework
agno>=0.1.0  # Agentic framework

# Kubernetes integration
kubernetes>=26.1.0

# Docker integration
docker>=6.1.1

# Messaging and caching
redis>=4.5.4
celery>=5.2.7
pika>=1.3.1

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.1.0

# Development tools
black>=23.3.0
isort>=5.12.0
flake8>=6.0.0
mypy>=1.2.0

# Monitoring and logging
prometheus-client>=0.16.0
python-json-logger>=2.0.7
opentelemetry-api>=1.15.0
opentelemetry-sdk>=1.15.0
opentelemetry-exporter-otlp>=1.15.0
